% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibodySeek.R
\name{antibodySeek}
\alias{antibodySeek}
\title{A function to run the antibodySeek shiny app}
\usage{
antibodySeek(...)
}
\arguments{
\item{...}{Any argument that you can pass to shiny::runApp}

\item{antibodies}{list of antibodies to search for}
}
\value{
Table antibodies and the fluorophores available in the Ansel Lab

Runs the antibodySeek shiny app.
}
\description{
This function runs the antibodySeek app
}
\examples{
# antibodySeek()
}
