% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panel.R
\name{get_panel}
\alias{get_panel}
\title{A function to create table of antibodies}
\usage{
get_panel(spreadsheet = NULL, antibodies = NULL)
}
\arguments{
\item{spreadsheet}{spreadsheet of antibodies}

\item{antibodies}{list of antibodies to search for}
}
\value{
Table antibodies and the fluorophores available in the Ansel Lab
}
\description{
This function takes a list of antibodies and returns a table of fluorophores
for each antibody that are available in the Ansel lab
}
\examples{
file <- antibodySeek::readData_example(path = "antibodies.csv")
spreadsheet <- readr::read_csv(file, col_names = TRUE)
spreadsheet <- tidyr::gather(spreadsheet,
Fluorophore, 
value, 
-Antibodies)
spreadsheet$Fluorophore <- factor(spreadsheet$Fluorophore,
levels = unique(spreadsheet$Fluorophore))
antibodies <- c("CD3", "CD4", "CD8", "Foxp3")
antibodySeek::get_panel(spreadsheet, antibodies)
}
